+++ lib/paginators/wallaby/abstract_resource_paginator.rb
@@ -12,21 +12,18 @@ module Wallaby
     end
 
     # Delegate methods to pagination provider
-    instance_methods =
-      ModelPaginationProvider.instance_methods - ::Object.instance_methods
-    delegate(*instance_methods, to: :@provider)
+    delegate(*(ModelPaginationProvider.instance_methods - ::Object.instance_methods), to: :@provider)
 
     # Paginator constructor
     # @param model_class [Class] model class
     # @param collection [#to_a] a collection of all the resources
     # @param params [ActionController::Parameters] parameters
     def initialize(model_class, collection, params)
-      @model_class = model_class || self.class.model_class
+      @model_class = self.class.model_class || model_class
       raise ArgumentError, 'model class required' unless @model_class
       @collection = collection
       @params = params
-      @provider =
-        Map.pagination_provider_map(@model_class).new(@collection, @params)
+      @provider = Map.pagination_provider_map(@model_class).new(@collection, @params)
     end
   end
 end